[Feat] 2차 재판(Debate) 도메인 분리
[Docs] 1차/2차 재판 로직 분리에 따른 스웨거 태그 및 API 경로 수정

[Refactor] Case/Debate 서비스 및 컨트롤러 책임 분리

- 기존 CaseController/Service에 혼재되어 있던 2차 재판(변론, 반론, 투표) 관련 로직을 신규 DebateController/Service로 분리
- CaseController/Service는 1차 재판(사건 생성, 초심 판결)의 책임만 갖도록 수정
- API 엔드포인트 일부 수정 (/second -> /debate)
- JudgmentEventListener가 CaseService 대신 DebateService를 참조하도록 변경하여 비동기 판결 로직이 올바르게 동작하도록 수정
