+++ src/main/java/com/demoday/ddangddangddang/controller/AuthController.java
@@ -22,7 +22,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = "Auth API", description = "사용자 인증 (회원가입, 로그인, 토큰 재발급) API")
+@Tag(name = "Auth API", description = "사용자 인증 (회원가입, 로그인, 토큰 재발급) API - by 최우혁")
 @RestController
 @RequestMapping("/api/v1/auth")
 @RequiredArgsConstructor
