+++ src/main/java/com/demoday/ddangddangddang/service/cases/DebateService.java
@@ -1,176 +1,41 @@
-package com.demoday.ddangddangddang.service;
+package com.demoday.ddangddangddang.service.cases;
 
 import com.demoday.ddangddangddang.domain.*;
 import com.demoday.ddangddangddang.domain.enums.*;
 import com.demoday.ddangddangddang.dto.ai.AiJudgmentDto;
-import com.demoday.ddangddangddang.dto.caseDto.*;
 import com.demoday.ddangddangddang.dto.caseDto.second.*;
 import com.demoday.ddangddangddang.global.code.GeneralErrorCode;
-import com.demoday.ddangddangddang.global.event.UpdateJudgmentEvent; // [추가] 이벤트
+import com.demoday.ddangddangddang.global.event.UpdateJudgmentEvent;
 import com.demoday.ddangddangddang.global.exception.GeneralException;
 import com.demoday.ddangddangddang.repository.*;
+import com.demoday.ddangddangddang.service.ChatGptService2;
 import com.demoday.ddangddangddang.service.ranking.RankingService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j; // [추가] log 사용을 위해
-import org.springframework.context.ApplicationEventPublisher; // [추가] 이벤트 발행
-import org.springframework.scheduling.annotation.Async; // [추가] 비동기 처리를 위해
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.function.Function;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class CaseService {
+public class DebateService {
 
     private final CaseRepository caseRepository;
-    private final ArgumentInitialRepository argumentInitialRepository;
     private final JudgmentRepository judgmentRepository;
-    private final ChatGptService2 chatGptService2;
-    private final DefenseRepository defenseRepository; // [추가] 2차 재판용
-    private final RebuttalRepository rebuttalRepository; // [추가] 2차 재판용
-    private final VoteRepository voteRepository; // [추가] 2차 재판용
-    private final LikeRepository likeRepository; // [추가] 2차 재판용
-    private final ApplicationEventPublisher eventPublisher; // [추가] 이벤트 발행기
+    private final DefenseRepository defenseRepository;
+    private final RebuttalRepository rebuttalRepository;
+    private final VoteRepository voteRepository;
+    private final LikeRepository likeRepository;
+    private final ApplicationEventPublisher eventPublisher;
     private final RankingService rankingService;
-    private final CaseParticipationRepository caseParticipationRepository;
-
-    // --- [ 1차 재판(초심) 관련 메서드 ] ---
-
-    @Transactional
-    public CaseResponseDto createCase(CaseRequestDto requestDto, User user) {
-        if (!requestDto.getMode().equals(CaseMode.SOLO)) {
-            throw new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "현재 솔로 모드만 지원합니다.");
-        }
-
-        // 1. 사건(Case) 생성 및 저장
-        Case newCase = Case.builder()
-                .mode(requestDto.getMode())
-                .title(requestDto.getTitle())
-                .status(CaseStatus.FIRST)
-                .build();
-        caseRepository.save(newCase);
-
-        //caseParticipate 저장
-        CaseParticipation caseParticipation = CaseParticipation.builder()
-                .user(user)
-                .aCase(newCase)
-                .result(CaseResult.PENDING)
-                .build();
-        caseParticipationRepository.save(caseParticipation);
-
-        // 2. A측 1차 입장문 생성 및 저장
-        ArgumentInitial argumentA = ArgumentInitial.builder()
-                .aCase(newCase)
-                .user(user)
-                .type(DebateSide.A)
-                .mainArgument(requestDto.getArgumentAMain())
-                .reasoning(requestDto.getArgumentAReasoning())
-                .build();
-        argumentInitialRepository.save(argumentA);
-
-        // 3. B측 1차 입장문 생성 및 저장
-        ArgumentInitial argumentB = ArgumentInitial.builder()
-                .aCase(newCase)
-                .user(user)
-                .type(DebateSide.B)
-                .mainArgument(requestDto.getArgumentBMain())
-                .reasoning(requestDto.getArgumentBReasoning())
-                .build();
-        argumentInitialRepository.save(argumentB);
-
-        // --- [ 4. AI 판결 로직 ] ---
-        // TODO: AI 호출이 오래 걸릴 수 있으므로, @Async 등을 이용한 비동기 처리로 리팩토링 권장
-        // 현재는 1차 재판 생성 시 바로 AI 판결을 받으므로 동기 처리
-        List<ArgumentInitial> arguments = List.of(argumentA, argumentB);
-        AiJudgmentDto aiResult = chatGptService2.getAiJudgment(newCase, arguments);
-
-        // 5. AI의 판결 결과를 DB에 저장
-        Judgment judgment = Judgment.builder()
-                .aCase(newCase)
-                .stage(JudgmentStage.INITIAL)
-                .content(aiResult.getVerdict()) // AI가 생성한 판결 내용
-                .basedOn(aiResult.getConclusion()) // AI가 생성한 결론
-                .ratioA(aiResult.getRatioA()) // AI가 생성한 비율
-                .ratioB(aiResult.getRatioB()) // AI가 생성한 비율
-                .build();
-        judgmentRepository.save(judgment);
-
-        return new CaseResponseDto(newCase.getId());
-    }
-
-    @Transactional(readOnly = true)
-    public JudgmentResponseDto getJudgment(Long caseId, User user) {
-        Case foundCase = findCaseById(caseId); // 헬퍼 메서드 사용
-
-        // TODO: 본인 확인 로직 (솔로 모드이므로 본인이 생성한 사건인지 확인)
-
-        // 초심(INITIAL) 판결문 조회
-        Judgment judgment = judgmentRepository.findByaCase_IdAndStage(caseId, JudgmentStage.INITIAL)
-                .orElseThrow(() -> new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "아직 판결이 완료되지 않았습니다."));
-
-        return new JudgmentResponseDto(judgment);
-    }
-
-    @Transactional(readOnly = true)
-    public CaseDetailResponseDto getCaseDetail(Long caseId, User user) {
-        Case aCase = findCaseById(caseId); // 헬퍼 메서드 사용
-
-        // 2. 1차 입장문 A, B 조회 (A, B 순서로 정렬)
-        List<ArgumentInitial> arguments = argumentInitialRepository.findByaCaseOrderByTypeAsc(aCase);
-
-        if (arguments.size() < 2) {
-            throw new GeneralException(GeneralErrorCode.INTERNAL_SERVER_ERROR, "데이터 오류: 입장문을 찾을 수 없습니다.");
-        }
-
-        ArgumentInitial argA = arguments.get(0);
-        ArgumentInitial argB = arguments.get(1);
-
-        // [추가] 보안 검증
-        if (aCase.getMode() == CaseMode.SOLO && !argA.getUser().getId().equals(user.getId())) {
-            throw new GeneralException(GeneralErrorCode.FORBIDDEN, "해당 사건에 대한 조회 권한이 없습니다.");
-        }
-        // TODO: (추후 확장) 파티 모드일 경우, 참여자인지 확인하는 로직 추가
-
-        rankingService.addCaseScore(caseId,1.0);
-
-        // 3. DTO 빌드
-        return CaseDetailResponseDto.builder()
-                .aCase(aCase)
-                .argA(argA)
-                .argB(argB)
-                .build();
-    }
-
-    // --- [ 1차 재판 상태 변경 (종료) ] ---
-    // (이 메서드는 2차 재판 시작 로직과 분리되어야 함)
-    @Transactional
-    public void updateCaseStatus(Long caseId, CaseStatusRequestDto requestDto, User user) {
-        Case foundCase = findCaseById(caseId);
-        // TODO: 본인 확인 로직
-
-        // 2차 재판 시작은 별도의 /appeal API를 이용하도록 명시
-        if (requestDto.getStatus() == CaseStatus.SECOND) {
-            throw new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "2차 재판 시작은 /appeal API를 이용해주세요.");
-        }
-
-        // DONE 상태만 허용 (2차 재판은 startAppeal 메서드로)
-        if (requestDto.getStatus() == CaseStatus.DONE) {
-            foundCase.updateStatus(CaseStatus.DONE);
-        } else {
-            throw new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "잘못된 상태 값입니다. DONE만 가능합니다.");
-        }
-    }
-
-
-    // --- [ 2차 재판 신규 메서드 ] ---
+    private final ChatGptService2 chatGptService2;
 
     /**
      * 2차 재판 시작 (시간제한 없음)
@@ -206,7 +71,7 @@ public class CaseService {
      * 2차 재판 상세 정보 조회
      */
     @Transactional(readOnly = true)
-    public CaseDetail2ndResponseDto getCaseDefenses(Long caseId, User user) {
+    public CaseDetail2ndResponseDto getDebateDetails(Long caseId, User user) {
         Case aCase = findCaseById(caseId);
         checkCaseStatusIsSecond(aCase); // 2차 재판 상태인지 확인
 
@@ -236,19 +101,13 @@ public class CaseService {
         Case aCase = findCaseById(caseId);
         checkCaseStatusIsSecond(aCase); // 2차 재판 상태인지 확인
 
-        // 1. 이 사건의 모든 변론 조회
         List<Defense> defenses = defenseRepository.findAllByaCase_Id(caseId);
-
-        // 2. 내가 좋아요 누른 변론 ID 목록 조회
         Set<Long> userLikedDefenseIds = likeRepository.findAllByUserAndContentType(user, ContentType.DEFENSE)
                 .stream().map(Like::getContentId).collect(Collectors.toSet());
-
-        // 3. 이 사건의 모든 반론을 가져와서, 변론 ID별로 개수(count)를 계산 (N+1 방지)
         Map<Long, Long> rebuttalCounts = rebuttalRepository.findAllByDefense_aCase_Id(caseId)
                 .stream()
                 .collect(Collectors.groupingBy(r -> r.getDefense().getId(), Collectors.counting()));
 
-        // 4. DTO로 변환
         return defenses.stream()
                 .map(defense -> DefenseResponseDto.fromEntity(
                         defense,
@@ -261,17 +120,11 @@ public class CaseService {
     /**
      * 반론 목록 조회
      */
-    // --- [ 3. (신규) 반론 목록 조회 API 로직 ] ---
     @Transactional(readOnly = true)
     public List<RebuttalResponseDto> getRebuttalsByDefense(Long defenseId, User user) {
-        // 1. 이 변론의 모든 반론 조회
         List<Rebuttal> rebuttals = rebuttalRepository.findAllByDefense_Id(defenseId);
-
-        // 2. 내가 좋아요 누른 반론 ID 목록 조회
         Set<Long> userLikedRebuttalIds = likeRepository.findAllByUserAndContentType(user, ContentType.REBUTTAL)
                 .stream().map(Like::getContentId).collect(Collectors.toSet());
-
-        // 3. DTO의 static 헬퍼 메서드를 사용해 중첩 구조(Tree)로 빌드
         return RebuttalResponseDto.buildTree(rebuttals, userLikedRebuttalIds);
     }
 
@@ -289,14 +142,10 @@ public class CaseService {
                 .type(requestDto.getSide())
                 .content(requestDto.getContent())
                 .build();
-
         Defense savedDefense = defenseRepository.save(defense);
 
-        rankingService.addCaseScore(caseId,5.0);
-
-        // AI 판결 업데이트 이벤트 발행 (비동기로 AI 판결을 요청)
+        rankingService.addCaseScore(caseId, 5.0);
         eventPublisher.publishEvent(new UpdateJudgmentEvent(caseId));
-
         return savedDefense;
     }
 
@@ -307,15 +156,12 @@ public class CaseService {
     public Rebuttal createRebuttal(RebuttalRequestDto requestDto, User user) {
         Defense defense = defenseRepository.findById(requestDto.getDefenseId())
                 .orElseThrow(() -> new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "원본 변론을 찾을 수 없습니다."));
-
-        checkCaseStatusIsSecond(defense.getACase()); // 2차 재판 상태인지 확인
+        checkCaseStatusIsSecond(defense.getACase());
 
         Rebuttal parentRebuttal = null;
         if (requestDto.getParentId() != null) {
             parentRebuttal = rebuttalRepository.findById(requestDto.getParentId())
                     .orElseThrow(() -> new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "부모 반론을 찾을 수 없습니다."));
-
-            // 부모 반론이 요청된 변론에 속하는지 확인
             if (!parentRebuttal.getDefense().getId().equals(defense.getId())) {
                 throw new GeneralException(GeneralErrorCode.INVALID_PARAMETER, "잘못된 부모 반론 ID 입니다. 해당 변론에 속하지 않습니다.");
             }
@@ -324,18 +170,14 @@ public class CaseService {
         Rebuttal rebuttal = Rebuttal.builder()
                 .defense(defense)
                 .user(user)
-                .parent(parentRebuttal) // 부모 반론 설정 (대댓글)
-                .type(requestDto.getType()) // A/B 입장
+                .parent(parentRebuttal)
+                .type(requestDto.getType())
                 .content(requestDto.getContent())
                 .build();
-
         Rebuttal savedRebuttal = rebuttalRepository.save(rebuttal);
 
-        rankingService.addCaseScore(defense.getACase().getId(),5.0);
-
-        // AI 판결 업데이트 이벤트 발행
+        rankingService.addCaseScore(defense.getACase().getId(), 5.0);
         eventPublisher.publishEvent(new UpdateJudgmentEvent(defense.getACase().getId()));
-
         return savedRebuttal;
     }
 
@@ -345,28 +187,21 @@ public class CaseService {
     @Transactional
     public void castVote(Long caseId, VoteRequestDto requestDto, User user) {
         Case aCase = findCaseById(caseId);
-        checkCaseStatusIsSecond(aCase); // 2차 재판 상태인지 확인
+        checkCaseStatusIsSecond(aCase);
 
         Vote vote = voteRepository.findByaCase_IdAndUser_Id(caseId, user.getId())
                 .map(existingVote -> {
-                    // 이미 투표했다면, 입장 변경
                     existingVote.updateChoice(requestDto.getChoice());
                     return existingVote;
                 })
-                .orElseGet(() -> {
-                    // 첫 투표라면, 새로 생성
-                    return Vote.builder()
-                            .aCase(aCase)
-                            .user(user)
-                            .type(requestDto.getChoice())
-                            .build();
-                });
-
+                .orElseGet(() -> Vote.builder()
+                        .aCase(aCase)
+                        .user(user)
+                        .type(requestDto.getChoice())
+                        .build());
         voteRepository.save(vote);
 
-        rankingService.addCaseScore(caseId,3.0);
-
-        // AI 판결 업데이트 이벤트 발행
+        rankingService.addCaseScore(caseId, 3.0);
         eventPublisher.publishEvent(new UpdateJudgmentEvent(caseId));
     }
 
@@ -391,50 +226,34 @@ public class CaseService {
                 .build();
     }
 
-
     /**
      * [비동기] 최종 AI 판결 업데이트 (이벤트 리스너가 호출)
-     * @Async 어노테이션을 통해 별도의 스레드에서 실행됩니다.
      */
     @Async
     @Transactional
     public void updateFinalJudgment(Long caseId) {
         log.info("Starting async judgment update for caseId: {}", caseId);
-
-        // ❗️ findCaseById가 이 클래스의 private 메서드가 아니라면,
-        //    caseRepository.findById를 사용해야 합니다.
         Case aCase = caseRepository.findById(caseId)
                 .orElseThrow(() -> new GeneralException(GeneralErrorCode.CASE_NOT_FOUND));
 
-        // 1. ⭐️ [수정] '채택된(adopted)' 변론 및 반론 조회
         List<Defense> adoptedDefenses = defenseRepository.findByaCase_IdAndIsAdopted(caseId, true);
         List<Rebuttal> adoptedRebuttals = rebuttalRepository.findAdoptedRebuttalsByCaseId(caseId);
-
-        // 2. 투표 결과 조회 (기존과 동일)
         long votesA = voteRepository.countByaCase_IdAndType(caseId, DebateSide.A);
         long votesB = voteRepository.countByaCase_IdAndType(caseId, DebateSide.B);
 
-        // 3. ⭐️ [수정] AI에게 판결 요청 (새로운 시그니처 사용)
         AiJudgmentDto aiResult = chatGptService2.requestFinalJudgment(
-                aCase,
-                adoptedDefenses,  // 'topDefenses' -> 'adoptedDefenses'
-                adoptedRebuttals, // 'adoptedRebuttals' 추가
-                votesA,
-                votesB
+                aCase, adoptedDefenses, adoptedRebuttals, votesA, votesB
         );
 
-        // 4. 'FINAL' 판결문 조회 및 업데이트 (기존과 동일)
         Judgment finalJudgment = judgmentRepository.findByaCase_IdAndStage(caseId, JudgmentStage.FINAL)
                 .orElseThrow(() -> new GeneralException(GeneralErrorCode.INTERNAL_SERVER_ERROR, "2차 판결문이 생성되지 않았습니다."));
 
-        // AI 결과로 판결문 내용 업데이트 (기존과 동일)
         finalJudgment.updateJudgment(
                 aiResult.getVerdict(),
                 aiResult.getConclusion(),
                 aiResult.getRatioA(),
                 aiResult.getRatioB()
         );
-        // judgmentRepository.save(finalJudgment); // @Transactional이므로 변경 감지로 자동 저장됩니다 (명시적 save도 무방)
         log.info("Finished async judgment update for caseId: {}", caseId);
     }
